requirements-practicalfinance.txt
Babel==2.9.0
py-moneyed==1.1
pytz==2020.5
